# optimize resizeNearestNeighborBackprop CPU