#### Description
<!--
Please describe the pull request here.
Also, if this is an issue/bug fix, please add the issue link for reference here.
-->
Optimizes the CPU implementation of `resizeNearestNeighborBackprop`

Benchmark code.  To run, expose resizeNearestNeighborBackprop at top level:
```
tf.setBackend("cpu");
    const ns = [64, 128, 256, 512];
    const RUNS = 100;

    ns.forEach(n => {
      const dy = tf.randomUniform([8, n, n, 3]);
      const x = tf.randomUniform([8, n / 2, n / 2, 3])
      
	// Warmup.
      tf.image.resizeNearestNeighborBackprop(dy, x, false);

      let res = null;
      const start = performance.now();
      for (let i = 0; i < RUNS; i++) {
        res = tf.image.resizeNearestNeighborBackprop(dy, x, false);
      }
      res.dataSync();
      const elapsed = (performance.now() - start) / RUNS;
      console.log(`N: ${n}: ${elapsed.toFixed(2)} ms`);
});
```
results before optimization:
```
N: 64: 18.59 ms
N: 128: 78.92 ms
N: 256: 312.69 ms
N: 512: 1290.73 ms
```

results after optimization:
```
N: 64: 3.07 ms
N: 128: 11.20 ms
N: 256: 43.60 ms
N: 512: 175.77 ms
```
---
<!-- Please do not delete this section -->
##### For repository owners only:

Please remember to apply all applicable tags to your pull request.
Tags: FEATURE, BREAKING, BUG, PERF, DEV, DOC, SECURITY

For more info see: https://github.com/tensorflow/tfjs/blob/master/DEVELOPMENT.md