@jgartman 2018-10-22
Oh, I just assumed that optimizing performance was something we'd want to do in general, I didn't have any motivation beyond that.  If you'd like I can hold off on opening any similar PR's until there's a specific use case for them.