@nsthorat 2018-10-22
No need to wait! Optimizations are always welcome, was just curious :)